% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.Get_Group.R
\name{call.Get_Group}
\alias{call.Get_Group}
\title{call.Get_Group}
\usage{
call.Get_Group(x, include_permissions = NULL, ...)
}
\arguments{
\item{x}{a group name or group id of length 1}

\item{include_permissions}{denote whether to also include permissions data for each group}

\item{...}{additional arguments to ApiRequest}
}
\value{
json
}
\description{
Base call to Permissions.GetGroup
}

