% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckExportJobStatus.R
\name{CheckExportJobStatus}
\alias{CheckExportJobStatus}
\title{Custom recursive status checker for Classifications export job requests}
\usage{
CheckExportJobStatus(id, iters = 0L, max_iters = 10L, iter.wait = 10)
}
\arguments{
\item{id}{(required)}

\item{iters}{(optional)}

\item{max_iters}{(optional)}

\item{iter.wait}{(optional)}
}
\value{
See \code{\link{Classifications_GetStatus}}
}
\description{
Periodically check the status of a job_id from a call to Classifications.CreateExport
}
\examples{
# TBD
}
\seealso{
Other Classifications methods: \code{\link{Classifications_CreateExport}},
  \code{\link{Classifications_GetElements}},
  \code{\link{Classifications_GetFilters}},
  \code{\link{Classifications_GetStatus}},
  \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}}
}
