% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk_shares.R
\name{chk_shares}
\alias{chk_shares}
\title{chk_shares}
\usage{
chk_shares(x)
}
\arguments{
\item{x}{A list of shares, possibly made with \code{\link{make_sharelist}}}
}
\value{
\code{TRUE} or an error (with error message)
}
\description{
Internal function - check input structure of 'shares' for a call to Segments.Get with shares
}
\details{
This performs the following checks, in order:

\enumerate{
\item High-level structure, by determining if \emph{x} is coercible to 
      a \code{data.frame}.
\item names, which must be \code{type,name}
\item values of field \code{type}, which must be one or both of \code{user,group}
\item column classes, which must be \code{scalar}, via \code{\link[jsonlite]{unbox}}
}

Failing any of these checks is an error.
Passing all of these checks returns \code{TRUE}
}
\examples{
#TBD
}
\seealso{
Other internal: \code{\link{collapse_simple_listcol}},
  \code{\link{recur_flatten_nested_container}},
  \code{\link{split_segment_ret}},
  \code{\link{split_single_row_ret}}
}

