% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifications_GetElements.R
\name{Classifications_GetElements}
\alias{Classifications_GetElements}
\title{Get report suite elements that are classification-enabled}
\usage{
Classifications_GetElements(rsid_list, preprocess_rsids = TRUE)
}
\arguments{
\item{rsid_list}{(Required) A character vector of report suite IDs to get compatibility elements for.}

\item{preprocess_rsids}{(Optional) If \code{TRUE}, uniquify case-normalized \emph{rsid_list} values, removing
\code{NA} and \code{""} as well.}
}
\value{
A nested \code{data.frame}, where \code{nrow(data.frame) == length(rsid_list)}, containing report suite id, name,
and a nested field, \emph{compatibility_elements}. The latter contains two fields named \code{id,name} denoting 
the classification-enabled element \code{id} and element \code{name}, respectively. See e.g. \code{\link[tidyr]{unnest}}
to flatten this field or \code{flatten_nested_df} in \code{wzAAHelper}.
}
\description{
Get the current classification-compatible elements for one or more report suites
}
\details{
This function calls the Adobe Analytics \code{Classifications.GetCompatibilityElements} method. The method supports
multiple report suites in a single call, with the condition that the every \emph{rsid_list} value has at least one
classification-enabled element. A single report suite where this is not the case will raise an error.

This may be a useful function to call prior to e.g. \code{\link{Classifications_GetTemplate}}, to get valid
\emph{element} values.
}
\note{
The API does not distinguish between valid and accessible (within your access credentials) report suites. Completely
invalid (i.e. non-existant), valid but inaccessible report suite IDs, and valid report suites that do not have at least
one classification-enabled element will raise the same server message:

\preformatted{ERROR: Bad Request  -  Access denied for the selected report suite:}
}
\examples{
\dontrun{
Classifications_GetElements(c("my_report_suite_1", "my_report_suite_2"))
}
}
\seealso{
Other Classifications methods: \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}}
}
