% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifications_GetExport.R
\name{Classifications_GetExport}
\alias{Classifications_GetExport}
\title{Get one or all pages from a complete export job}
\usage{
Classifications_GetExport(status_return = NULL, all_pages = TRUE,
  page = NULL)
}
\arguments{
\item{status_return}{(required) The return from a complete export job}

\item{all_pages}{(optional) Should all available pages be returned? Defaults to \code{TRUE}}

\item{page}{(optional) If provided, will only return this specific page. Ignored if \code{all_pages} is \code{TRUE}}
}
\value{
A named, nested \code{list} with as many elements as the number of pages
returned, for each \emph{file_id}. The (Each) first-level element will
be of length 3, with the following basic structure:

 \itemize{
 \item \code{warnings} A \code{logical} vector of length 1. \code{NA}
       if no warnings.
 \item \code{header} A \code{list} of length 1, the sub-element containing 
       the field names for the returned \code{data}
 \item \code{data} A \code{list} of length 1, the sub-element containing a
       nested \code{data.frame}. 
 } 
 
Each element is named using the \code{job_id}, \code{file_id}, and page number of the input
\emph{status_return}.
}
\description{
Get one or all pages from completed classification job export
}
\details{
\href{https://marketing.adobe.com/developer/documentation/classifications-1-4-saint/r-getexport}{Adobe 1.4 documentation}

The extends the basic functionality of the baseline method by allowing all possible \code{file_id} pages to be 
returned (default). Set the value of \emph{all_pages} to \code{FALSE} to reproduce the behavior of the 
default method.
}
\note{
The returned value is nested one additional value to ensure a consistent structure for single or multi-page
returns.
}
\section{Access Privileges}{
 
This function calls an Adobe Analytics method that requires administrative/elevated privileges for
some or all functionality.
}

\examples{
# TBD
}
\seealso{
Other Classifications methods: \code{\link{CheckExportJobStatus}},
  \code{\link{Classifications_CreateExport}},
  \code{\link{Classifications_GetElements}},
  \code{\link{Classifications_GetFilters}},
  \code{\link{Classifications_GetStatus}},
  \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}}
}
