% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifications_GetFilters.R
\name{Classifications_GetFilters}
\alias{Classifications_GetFilters}
\title{Get valid date and classification element fields for filtering}
\usage{
Classifications_GetFilters(element = NULL, rsid_list = NULL)
}
\arguments{
\item{element}{(required) The target element (id). Must be a vector of length 1}

\item{rsid_list}{(required) One or more report suites.}
}
\value{
A \code{list} of length 2, containing \code{columns} and \code{filter_dates}. 
This holds even when more than one report suite is requested.
}
\description{
Figure out valid date and classification column filters for an element and report suite
}
\details{
\href{https://marketing.adobe.com/developer/documentation/classifications-1-4-saint/r-getfilters-1}{Adobe 1.4 documentation}:  

This is a useful ancillary function to obtain filter parameters for use with 
e.g. \code{\link{Classifications_CreateExport}}.
}
\note{
It is assumed that when more than one report suite is called, the results
represent the intersection of values, although this remains to be confirmed; it
is entirely possible that the results could instead be the union. Or something
else altogether.

At the moment, this function provides no error checking. This will be updated
in the future.
}
\section{Access Privileges}{
 
This function calls an Adobe Analytics method that requires administrative/elevated privileges for
some or all functionality.
}

\examples{
\dontrun{
Classifications_GetFilters("evar16", "my_rsid")
}
}
\seealso{
Other Classifications methods: \code{\link{CheckExportJobStatus}},
  \code{\link{Classifications_CreateExport}},
  \code{\link{Classifications_GetElements}},
  \code{\link{Classifications_GetExport}},
  \code{\link{Classifications_GetStatus}},
  \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}}
}
