% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifications_GetStatus.R
\name{Classifications_GetStatus}
\alias{Classifications_GetStatus}
\title{Get the status of a Classification CreateExport request}
\usage{
Classifications_GetStatus(id_or_ret = NULL)
}
\arguments{
\item{id_or_ret}{(required) An integer job_id of length 1, or the return from a 
call to \code{Classifications.CreateExport}}
}
\value{
A \code{data.frame} containing four fields:

\itemize{
\item \code{id} The \code{id} for each \code{type}
\item \code{type} The type of \code{id} returned. One of \code{job_id, file_id}
\item \code{viewable_pages} The number of result pages. For \code{type==file_id}, should be non-zero if
       results are ready, or available (possibly at all)
\item \code{status} The status for each \code{type}. A finished return should have key-value pairs of
      \code{job_id:Completed} and \code{file_id:Ready}. Other status codes are possible, depending 
      on the status at time of query.
}

In almost all cases, the return will have two rows, although it is possible to have single-row
returns during initial processing stages.
}
\description{
Get the status of a submitted request to Classifications_CreateExport
}
\details{
\href{https://marketing.adobe.com/developer/documentation/classifications-1-4-saint/r-getstatus-1}{Adobe 1.4 documentation}    

This function is required to obtain the \code{file_id} and number of pages, for a 
return from Classifications.CreateExport. The expected input can be either an 
\code{integer} id of length 1, or the direct return from 
\code{\link{Classifications_CreateExport}}.
}
\section{Access Privileges}{
 
This function calls an Adobe Analytics method that requires administrative/elevated privileges for
some or all functionality.
}

\examples{
\dontrun{
# Simplest case: pass the job_id directly
Classifications_GetStatus(my_integer_id)

# More realistic: Assume you have a job_id return from this call:
my_export_request <- 
Classifications_CreateExport(
    date_filter_start_date = "Jan 2017", 
    date_filter_end_date = "Nov 2017", 
    element = "evar16", 
    email_address = "my_email@xyz.com", 
    rsid_list = list("my_rsid_1")
    )
## Now pass the value explicitly:
Classifications_GetStatus(my_export_request$job_id)
## Or, simply pass the entire return
Classifications_GetStatus(my_export_request)
}
}
\seealso{
Other Classifications methods: \code{\link{CheckExportJobStatus}},
  \code{\link{Classifications_CreateExport}},
  \code{\link{Classifications_GetElements}},
  \code{\link{Classifications_GetExport}},
  \code{\link{Classifications_GetFilters}},
  \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}},
  \code{\link{parse_job_return}},
  \code{\link{split_multi_jobfile}}
}
\concept{Classifications methods}
