% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifications_GetTemplate.R
\name{Classifications_GetTemplate}
\alias{Classifications_GetTemplate}
\title{Get a SAINT import template}
\usage{
Classifications_GetTemplate(rsid_list = NULL, element = NULL,
  classification_names = NULL, encoding = "UTF-8")
}
\arguments{
\item{rsid_list}{(Required) A character vector of report suite IDs to get templates for.}

\item{element}{(Required) A character vector of length 1, or of any length but containing a single
unique value. This is the parent variable in Adobe Analytics that you wish to classify. See details.}

\item{classification_names}{(Optional) A character vector of the classification names to include in the template.
Will return all names by default.}

\item{encoding}{(Optional) The text encoding to use for the output. Defaults to \code{UTF-8} and generally should
be left as such. If provided, must be a character vector of length 1.}
}
\value{
A \code{data.frame}, where \code{nrow(data.frame) == length(rsid_list)}, containing metadata and a field 
called \emph{template}. Additionally, the requested \code{element} is captured as an attribute called 
\emph{classification_element}.

The \emph{template} field can be written out as a tab-separated file, or quickly viewed via \code{cat}.
}
\description{
Get a template for a single element for one or more report suites
}
\details{
This calls the \code{Classifications.GetTemplate} method. The method supports multiple report suites in a single
call, with the condition that the specified \emph{element} is available for classification for all requested 
report suites. A single report suite where the element is invalid will raise an error.

All values passed to \emph{rsid_list} are case-normalized and uniquified, with \code{NA} and \code{""} dropped, 
if present.
}
\section{Access Privileges}{
 
This function calls an Adobe Analytics method that requires administrative/elevated privileges for
some or all functionality.
}

\examples{
\dontrun{
Classifications_GetTemplate(c("my_suite_prod", "my_suite_dev"), element = "product")
}
}
