% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_simple_target.R
\name{collapse_simple_target}
\alias{collapse_simple_target}
\title{Collapse a target list-column to a delimited atomic vector}
\usage{
collapse_simple_target(x, targ = NULL)
}
\arguments{
\item{x}{A data.frame}

\item{targ}{A character vector of length 1 to denote the column in \emph{x} to operate on}
}
\value{
A unnested character vector; see \code{collapse_simple_listcol}.
}
\description{
A custom mechanism to unnest simple list-columns by collapsing into a delimited atomic vector
}
\details{
This wraps the internal function \code{\link{collapse_simple_listcol}} to handle more (and) typical
inputs, with some additional constraints. \emph{targ} is set to a default of \code{NULL}
intentionally, for flexibility and to enforce explicit provision of a value. However, it 
is expected that the valid values should only be:

\itemize{
\item \preformatted{tags}
\item \preformatted{compatibility}
}
}
\examples{

\dontrun{
collapse_tags   <- collapse_simple_target(x, targ = "tags")
collapse_compat <- collapse_simple_target(x, targ = "compatibility")
}
}

