% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_tags.R
\name{collapse_tags}
\alias{collapse_tags}
\title{collapse_tags}
\usage{
collapse_tags(x)
}
\arguments{
\item{x}{An unnested list of length 1}
}
\value{
A collapsed atomic vector, with each element of \code{x[[1]]}
separated by \code{, } (comma\code{space}). If a comma is encountered
in any element of \code{x[[1]]}, ALL elements will additionally be 
encapsulated by square brackets as a more readable proxy for 
quotes.
}
\description{
collapse nested tags list-column into a flattened atomic vector
}
\details{
Will likely not be exported for release, but being exported for development. 
This does not check for any upsteam (e.g. data.frame) names, and should be used
directly on a column accessed by name (or double-subscripted via index/name); can
be easily integrated into a pipeline, so this design is intentional in anticipation.
}
\examples{
# TBD
}

