% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten_container.R
\name{flatten_container}
\alias{flatten_container}
\title{flatten_container}
\usage{
flatten_container(x)
}
\arguments{
\item{x}{A single-row data.frame, generally the output from \code{\link{split_segment_ret}}}
}
\value{
A flattened segment definition
}
\description{
Mainly for debugging and development purposes; flattens segment definitions containing nested containers
}
\note{
This is temporarily being exported during development to facilitate abstraction. Might
be useful to export ultimately, with some tweaks. 

I suspect if this is to be exported, it should probably also bring in \code{\link{split_segment_ret}} and 
handle multi-row inputs using e.g. \code{Map()} or \code{lapply()}. 

At the moment, all this is doing is wrapping around the two (not-exported) functions:

\enumerate{
\item \code{\link{split_single_row_ret}}
\item \code{\link{recur_flatten_nested_container}}
}
}
\examples{
#TBD
}

