% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get_Segments.R
\name{Get_Segments}
\alias{Get_Segments}
\title{Get Adobe Analytics Segments information}
\usage{
Get_Segments(accessLevel = NULL, fields = NULL, selected = NULL,
  sort = NULL, filters = NULL, handle_tagsCol = FALSE, ...)
}
\arguments{
\item{accessLevel}{A character vector of length 1. Must be one of \code{all}, \code{shared}, or 
\code{owned}. If not specified, defaults to \code{owned}.}

\item{fields}{A character vector denoting the quantity, depth, and general detail of information desired. 
See details for permissible values. The API always includes \code{id} and \code{name}, by default.}

\item{selected}{A character vector of segment ID(s) you wish to query for. If both \code{selected} and 
\code{accessLevel} are provided, \code{selected} take precedence.}

\item{sort}{A character vector}

\item{filters}{A character vector}

\item{handle_tagsCol}{A logical vector of length 1}

\item{...}{Additional args to pass to \code{ApiRequest}}
}
\value{
A data.frame if successful.
}
\description{
Query the AA segments API to return segment information based on different criteria, at 
the level of desired detail
}
\examples{
# Forthcoming
}

