% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segment_body.R
\name{make_segment_body}
\alias{make_segment_body}
\title{Make a complete request body for creating or editing a segment}
\usage{
make_segment_body(segment_meta, segment_container)
}
\arguments{
\item{segment_meta}{A named list; metadata for container. Generally the output from \code{\link{make_segment_meta}}}

\item{segment_container}{A nested named list; container with one or more rules. Generally the output from 
\code{\link{make_segment_container}}}
}
\value{
A nested named list, generally passed to \code{\link{Save_Segment}}
}
\description{
Create a segment body to pass to Save_Segment()
}
\details{
This builds up the final required structure, including nesting \emph{segment_container} under a value of
 \code{definition}, as required by the API.
}
\note{
Premliminary checks in place for segment_meta, not yet for segment_container. Additional checks to be 
added for both.
}
\examples{
# Forthcoming
}
\seealso{
Other make_segment functions: \code{\link{make_element_rules}},
  \code{\link{make_segment_container}},
  \code{\link{make_segment_meta}},
  \code{\link{make_sharelist}}
}

