% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recur_flatten_nested_container.R
\name{recur_flatten_nested_container}
\alias{recur_flatten_nested_container}
\title{recur_flatten_nested_container}
\usage{
recur_flatten_nested_container(x, lst = c(), d = 0L)
}
\arguments{
\item{x}{The output of \code{\link{split_single_row_ret}}. Must contain an element named \code{segment_def}, 
which is output by \code{split_single_row_ret()}.}

\item{lst}{Used only for recursion; accumulates results}

\item{d}{Used only for recursion; used to name outputs and as internal index}
}
\value{
A flattened version of containers with names of definitions (now flattened)
 prefixed by \code{L[0-9]+}
}
\description{
Internal function - recursively flattens nested containers (definitions)
}
\note{
This function does not do much for simple, i.e. unnested segments, but is required
for nested segments, which includes stacked segments. 

This currently does NOT work for segments that have been created sub-optimally. The 
definition of \emph{sub-optimally} is subjective, but includes:

\itemize{
\item Unnecessarily nested containers
\item The above, interspersed with rules at varying hierarchies
\item Stacked segments where one or more upstream segments have been created
      in accordance with (but not limited to) the aforementioned possibilities
}

This function includes a helper that may not be required any longer, but does not hurt
}
\seealso{
Other internal: \code{\link{call.Get_base}},
  \code{\link{chk_shares}},
  \code{\link{collapse_simple_listcol}},
  \code{\link{split_segment_ret}},
  \code{\link{split_single_row_ret}}
}
