% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr_permissions.R
\name{restr_permissions}
\alias{restr_permissions}
\title{Restructure permissions from call.Get_Group}
\usage{
restr_permissions(x)
}
\arguments{
\item{x}{non-recursively unlisted result from call.Get_Group; see details}
}
\value{
a data.table or NULL
}
\description{
Function to handle permissions call return from call.Get_Group, permission = TRUE
}
\details{
Normally used in the context of restr.Get_Group. Exported for flexibility.
}
\examples{
#TBD
}
\seealso{
Other get group functions: \code{\link{call.Get_Group}},
  \code{\link{restr.Get_Group}}
}
