% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Segment.R
\name{Save_Segment}
\alias{Save_Segment}
\title{Create a new segment via Segments.Save}
\usage{
Save_Segment(x, ...)
}
\arguments{
\item{x}{A segment body, generally constructed with \code{make_segment_body()}}

\item{...}{Additional methods to pass to \code{ApiRequest()}.}
}
\value{
If successful, a data.frame with the name, report suite ID, a timestamp, the method, 
and the new segment_id.

Timestamp is always returned as a character-formatted vector of length 1, with timezone set to "UTC"
}
\description{
Push a new segment to a target report suite via the Adobe Analytics Segments.Save method
}
\note{
Error checking not yet implemented, but will be. At the moment, only errors if an
element named \code{id} is detected, as such elements are used to edit/modify, but
not create (brand-new) segments. 

The error message in this case refers to an EDIT version of this function, which needs
to be written.
}
\examples{
# Forthcoming
}

