% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveInternalURLFilters.R
\name{SaveInternalURLFilters}
\alias{SaveInternalURLFilters}
\title{SaveInternalURLFilters}
\usage{
SaveInternalURLFilters(rsid, urls)
}
\arguments{
\item{rsid}{A vector of report suite IDs}

\item{urls}{A vector of urls to filter, for the target report suite(s)}
}
\value{
\code{TRUE}, if successful
}
\description{
Save Internal URL Filters for one or more Report Suites
}
\details{
This is a convenient way to batch-add URLs to one or more report suites.
}
\note{
There is minimal input checking at the moment; this will be addressed prior to release, but 
use cautiously at the moment, and please raise an issue if you encounter any strange behavior.
}
\section{Access Privileges}{
 
This function calls an Adobe Analytics method that requires administrative/elevated privileges
}

\examples{
\dontrun{
my_url <- c("test_url_1_DOT_COM", "test_url_2.com")
my_rs  <- "my_reportSuiteID"

SaveInternalURLFilters(rsid = my_rs, urls = my_url)
}
}
