% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_segment_ret.R
\name{split_segment_ret}
\alias{split_segment_ret}
\title{split_segment_ret}
\usage{
split_segment_ret(seg_ret)
}
\arguments{
\item{seg_ret}{A data.frame from a successful call to Segments.Get via \code{\link{Segments_Get}}}
}
\value{
A list of length equal to \code{nrow(x)}
}
\description{
Internal function - splits data.frame into rows, with checks
}
\seealso{
Other internal: \code{\link{call.Get_base}},
  \code{\link{chk_shares}},
  \code{\link{collapse_simple_listcol}},
  \code{\link{collapse_simple_target}}
}
