% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_single_row_ret.R
\name{split_single_row_ret}
\alias{split_single_row_ret}
\title{split_single_row_ret}
\usage{
split_single_row_ret(seg_ret_single)
}
\arguments{
\item{seg_ret_single}{A single row from a split data.frame, i.e. the output of \code{\link{split_segment_ret}}}
}
\value{
A list of length two, with elements named \code{segment_meta} and \code{segment_def} containing,
repectively, the segment metadata (non-definition) and segment definition (\code{definition})
}
\description{
Internal function - Further splits and checks output from split_segment_ret
}
\note{
This is mainly a helper for \code{\link{recur_flatten_nested_container}}
}
\seealso{
Other internal: \code{\link{chk_shares}},
  \code{\link{collapse_simple_listcol}},
  \code{\link{recur_flatten_nested_container}},
  \code{\link{split_segment_ret}}
}
