% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call.Get_Segments.R
\name{call.Get_Segments}
\alias{call.Get_Segments}
\title{Get Adobe Analytics Segments information}
\usage{
call.Get_Segments(accessLevel = NULL, fields = NULL, selected = NULL,
  sort = NULL, filters = NULL, ...)
}
\arguments{
\item{accessLevel}{(optional) A character vector of length 1. Must be one of \code{all, shared, owned}. 
If not specified, defaults to \code{owned}.}

\item{fields}{(optional) A character vector denoting the quantity, depth, and general detail of information desired. 
Must be one of 
\code{tags, shares, description, owner, modified, compatibility, favorite, reportSuiteID, definition}. 
The API always includes \code{id} and \code{name}, by default.}

\item{selected}{(optional) A character vector of segment ID(s) you wish to query for. If both \code{selected} and 
\code{accessLevel} are provided, \code{selected} take precedence.}

\item{sort}{(optional) A character vector of length 1. Must be one of \code{id, name, description, reportSuiteID,
owner, modified, favorite}. If not specified, defaults to \code{id}.}

\item{filters}{(optional) A named \code{list}.}

\item{...}{Additional args to pass to \code{ApiRequest}}
}
\value{
A data.frame, possibly with nested columns depending on requested parameters within \emph{fields}. Notably, 
the following values in \emph{fields} return list-columns of varying complexity:

\itemize{
\item{tags}
\item{shares}
\item{compatibility}
\item{definition}
}

The number of rows corresponds to the number of unique segments, identified by the \code{id} field. With default
settings, a successful return will contain two fields, \code{id} and \code{name}.
}
\description{
Query the AA segments API to return segment information based on different criteria, at 
the level of desired detail
}
\note{
It is expected that once the full method is completed, this function will no longer be required by itself, 
although likely will still be exported for flexibility and debugging.
}
\examples{
# Forthcoming
}
\seealso{
Other get segments functions: \code{\link{restr.Get_Segments}}
}

