% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckExportJobStatus.R
\name{CheckExportJobStatus}
\alias{CheckExportJobStatus}
\title{Custom recursive status checker for Classifications export job requests}
\usage{
CheckExportJobStatus(id_or_ret, max_iters = 10L, iter.wait = 10,
  iters = 0L)
}
\arguments{
\item{id_or_ret}{(required) An integer job_id of length 1, or the return from a 
call to Classifications.CreateExport}

\item{max_iters}{(optional) Total number of iterations to call before returning the 
current status. Defaults to \code{10L}}

\item{iter.wait}{(optional) Number of seconds to wait between each successive attempt.
Defaults to \code{10L}}

\item{iters}{Internal iteration accumulator}
}
\value{
See \code{\link{Classifications_GetStatus}} for the return upon a 
non-error termination. Furthermore, a message denoting the current status
for each attempt.
}
\description{
Periodically check the status of a job_id from a call to Classifications.CreateExport
}
\details{
This is a recursive version of \code{\link{Classifications_GetStatus}}, and will
continue to check the status of a job export request until the job is ready, or the
number of iterations reaches \code{max_iters}, whichever happens first.
}
\note{
Currently, this function does not distinguish between zero-page returns that are technically
complete, but invalid, and valid requests that are still processing. 
This may be addresed in the near future.
}
\examples{
# Same as \\code{\\link{Classifications_GetStatus}}, 
#  except for the function called.
}
\seealso{
Other Classifications methods: \code{\link{Classifications_CreateExport}},
  \code{\link{Classifications_GetElements}},
  \code{\link{Classifications_GetExport}},
  \code{\link{Classifications_GetFilters}},
  \code{\link{Classifications_GetStatus}},
  \code{\link{Classifications_GetTemplate}},
  \code{\link{WriteTemplate}}
}
