% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_element_rules.R
\name{make_element_rules}
\alias{make_element_rules}
\title{Expand a vector of rules into scalar list-sets}
\usage{
make_element_rules(element, operator, rules, how = NULL)
}
\arguments{
\item{element}{A character vector of length 1. Must be a valid element (id) within access rights of your account}

\item{operator}{A character vector of length 1. For elements-based rules, the API defines allowable values. See details.}

\item{rules}{A character vector; if not, will be coerced via \code{as.character}}

\item{how}{Not used at the moment.}
}
\value{
A list of \code{length(rules)}, where length is calcualted after trimming leading and trailing whitespace and
de-duplicating \emph{rules} via \code{str_trim} and \code{unique}, respectively.
}
\description{
For the most common single element, single operator, many rules use case
}
\examples{
#TBD
}
\seealso{
Other make_segment functions: \code{\link{make_segment_body}},
  \code{\link{make_segment_container}},
  \code{\link{make_segment_meta}}
}

