% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segment_container.R
\name{make_segment_container}
\alias{make_segment_container}
\title{Make a segment container, including rules}
\usage{
make_segment_container(name = NULL, type = NULL, operator = NULL,
  rules = NULL, exclude = NULL)
}
\arguments{
\item{name}{(Optional) Container name. A character vector of length 1}

\item{type}{(Required) Container type. One of \code{hits,visits,visitors}. A character vector of length 1. Note
restrictions apply for sequential segments. Does not yet support a value of \code{logicgroup} for 
sub-containers.}

\item{operator}{(Optional) How should rules within the container be evaluated? One of \code{and,or,then}. A
character vector of length 1. Note that restrictions apply for \code{then}. If not provided, defaults to \code{and}. 
Not meaningful for single-rule containers.}

\item{rules}{(Required) A list of rules, with required names. Generally created with \code{make_element_rules}.}

\item{exclude}{(Optional) Should the container exclude, rather than include, data that matches the 
rule(s)? A logical vector of length 1. Defaults to \code{FALSE}}
}
\value{
A nested list
}
\description{
Construct a segment container (not definition) with one or more rules. Does not yet support sequential segments.
}
\examples{
# Forthcoming
}
\seealso{
Other make_segment functions: \code{\link{make_element_rules}},
  \code{\link{make_segment_body}},
  \code{\link{make_segment_meta}}
}

