% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_segment_meta.R
\name{make_segment_meta}
\alias{make_segment_meta}
\title{Make segment-level metadata for creating (NOT editing) a segment}
\usage{
make_segment_meta(name = NULL, reportSuiteID = NULL, description = NULL,
  favorite = NULL, owner = NULL, shares = NULL, tags = NULL)
}
\arguments{
\item{name}{(Required) A character vector of length 1. This is the name (title) of the segment}

\item{reportSuiteID}{(Required) A character vector of length 1. This is the report suite the segment should
belong to. Must be a valid report suite ID and one that your login has permissions to access}

\item{description}{(Optional) A character vector of length 1. The segment description}

\item{favorite}{(Optional) A logical vector of length 1. Should the segment be added to the favorite segments list?}

\item{owner}{(Optional) A character vector of length 1. The Login of the user who will be the owner. Defaults to 
the current user if not provided.}

\item{shares}{NOT CURRENTLY SUPPORTED; when this changes: 
(Optional) A named list of length 2 with names \emph{type} (one of \code{group,user}) and \code{name}, 
which is the group or user name to share with. Both values must be provided as character vectors of length 1.}

\item{tags}{NOT CURRENTLY SUPPORTED; when this changes: 
(Optional) A character vector of keywords to group segments for filtering}
}
\value{
A list, possibly nested, containing the required segment-level metadata
}
\description{
For a new segment, create required metadata and output appropriate structure for a call to Get_Segments
}
\details{
WORK IN PROGRESS; this creates the appropriate segment-level metadata, required as a part of creating
a new segment. 

The scope of this function is for creating a new segment via the Adobe Analytics \code{Segments.Save}
method, specifically for creating a NEW segment, as opposed to editing (updating) an existing segment. 

This separation is intentional, as this function will eventually be integrated into the final exported function
\code{Save_Segment()}, which does not yet exist. This means that this particular function will likely NOT be 
exported in the public release.
}
\examples{
# Forthcoming
}
\seealso{
Other make_segment functions: \code{\link{make_element_rules}},
  \code{\link{make_segment_body}},
  \code{\link{make_segment_container}}
}

