% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.Get_Group.R
\name{restr.Get_Group}
\alias{restr.Get_Group}
\title{Create user-friendly output from call.Get_Group return}
\usage{
restr.Get_Group(call_ret)
}
\arguments{
\item{call_ret}{result from call.Get_Group}
}
\value{
list of data.tables
}
\description{
Handle all return possibilities from call.Get_Group, with or without permissions
}
\examples{
#TBD
}
\seealso{
Other get.group.functions: \code{\link{call.Get_Group}},
  \code{\link{restr_permissions}}
}

