% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.Get_Group.R
\name{restr.Get_Group}
\alias{restr.Get_Group}
\title{Restructure json return from call.Get_Group}
\usage{
restr.Get_Group(call_ret)
}
\arguments{
\item{call_ret}{result from call.Get_Group}
}
\value{
list of data.tables
}
\description{
Restructure json return from call.Get_Group
}
\examples{
#TBD
}

