% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restr.Get_Segments.R
\name{restr.Get_Segments}
\alias{restr.Get_Segments}
\title{restr.Get_Segments}
\usage{
restr.Get_Segments(x, collapse_rules = FALSE, bind_rules = FALSE)
}
\arguments{
\item{x}{A call return from \code{Get_Segments}}

\item{collapse_rules}{logi. Should the output collapse multi-ruled into comma-delimited format? 
Defaults to \code{FALSE}}

\item{bind_rules}{logi. Should the output return rules as individual lists of \code{data.table}s or 
attempt to glue them together via \code{rbindlist}? Defaults to \code{FALSE}}
}
\value{
Depending on the values passed to \emph{collapse_rules} and \emph{bind_rules}, a list of outputs of
varying lengths
}
\description{
Restructure the return from a call to Segments.Get, via Get_Segments. Currently does NOT work
properly for nested containers or stacked segments! Much more work required, but this works 
for the most basic cases.
}
\note{
Default args for \emph{collapse_rules} and \emph{bind_rules} are not the most useful by far, but
are the safest in the sense that they will work for nested containers and stacked segments, although
the restructured output for these kinds of segments is hardly ideal. 

This is definitely not a complete function yet; you have been warned.
}
\examples{
#Forthcoming
}

