% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Save_Segment.R
\name{Save_Segment}
\alias{Save_Segment}
\title{Create a new segment via Segments.Save}
\usage{
Save_Segment(x, ..., override_and_edit = FALSE)
}
\arguments{
\item{x}{A segment body. Helper: \code{\link{make_segment_body}}}

\item{...}{Additional methods to pass to \code{\link[RSiteCatalyst]{ApiRequest}}}

\item{override_and_edit}{logical value to override safety check and use this function for editing. 
\code{FALSE} by default. Only works if a value for \emph{id} is present}
}
\value{
If successful, a data.frame with the name, report suite ID, a timestamp, the method, 
and the new segment_id.

Timestamp is always returned as a character-formatted vector of length 1, with timezone set to "UTC"
}
\description{
Push a new segment to a target report suite via the Adobe Analytics Segments.Save method
}
\note{
Errors if an element named \code{id} is detected, as such elements are used to edit/modify, but
not create (brand-new) segments. The error message in this case refers to an EDIT version of this function, 
which needs to be written. 

IMPORTANT: For temporary convenience only, this function has an override so it can be used
to edit, while the edit method is written. Since there are no checks for EDIT, this is an 
unchecked hack to edit. Since this is temporary, there is no check for \emph{id} being
the required scalar type (via \code{\link[jsonlite]{unbox}}).
}
\examples{
# Forthcoming
}

